requirements.txt
django>=4.2
django-environ
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-jazzmin
celery
redis
requests
pillow
gunicorn
whitenoise
psycopg2-binary
google-genai
django-extensions
drf-yasg
transformers
torch
torchaudio
soundfile
